+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,72 @@
 # @ballerine/workflows-service
 
+## 0.7.55
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.52
+  - @ballerine/common@0.9.40
+  - @ballerine/workflow-node-sdk@0.6.52
+
+## 0.7.54
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.51
+  - @ballerine/workflow-node-sdk@0.6.51
+  - @ballerine/common@0.9.39
+
+## 0.7.53
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.38
+  - @ballerine/workflow-core@0.6.50
+  - @ballerine/workflow-node-sdk@0.6.50
+
+## 0.7.52
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.37
+  - @ballerine/workflow-core@0.6.49
+  - @ballerine/workflow-node-sdk@0.6.49
+
+## 0.7.51
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.36
+  - @ballerine/workflow-core@0.6.48
+  - @ballerine/workflow-node-sdk@0.6.48
+
+## 0.7.50
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.35
+  - @ballerine/workflow-core@0.6.47
+  - @ballerine/workflow-node-sdk@0.6.47
+
+## 0.7.49
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.34
+  - @ballerine/workflow-core@0.6.46
+  - @ballerine/workflow-node-sdk@0.6.46
+
 ## 0.7.48
 
 ### Patch Changes